Add LoadingBackdrop component to show a loading indicator
This commit introduces the  component which renders a semi-transparent background with a circular loading indicator. This component is used to indicate that a background task, such as downloading data, is being carried out.

The component uses Material-UI's Backdrop and CircularProgress to achieve a clean and consistent layout. Additionally, this component has been reused in various parts of the application to improve consistency and reduce code duplication.

Additionally, several unnecessary imports in various files have been removed to improve code cleanliness and efficiency.